backend/requirements.txt
# Backend requirements
Flask
python-dotenv
pyodbc
gunicorn
